+++ app/model/recordings.js
@@ -26,6 +26,7 @@ var dbpool       = require('../utils/dbpool');
 var tyler        = require('../utils/tyler.js');
 
 const moment = require('moment');
+const Projects = require('./projects');
 
 // local variables
 var s3, s3RFCx;
@@ -69,11 +70,11 @@ var Recordings = {
     QUERY_FIELDS : {
         id     : { subject: 'R.recording_id',     project:  true },
         site   : { subject: 'S.name',             project: false, level:1, next: 'year'                },
-        year   : { subject: 'YEAR(R.datetime)',   project: true,  level:2, next: 'month' , prev:'site' },
-        month  : { subject: 'MONTH(R.datetime)',  project: true,  level:3, next: 'day'   , prev:'year' },
-        day    : { subject: 'DAY(R.datetime)',    project: true,  level:4, next: 'hour'  , prev:'month'},
-        hour   : { subject: 'HOUR(R.datetime)',   project: true,  level:5, next: 'minute', prev:'day'  },
-        minute : { subject: 'MINUTE(R.datetime)', project: true,  level:6,                 prev:'hour' }
+        year   : { subject: 'YEAR(R.datetime_local)',   project: true,  level:2, next: 'month' , prev:'site' },
+        month  : { subject: 'MONTH(R.datetime_local)',  project: true,  level:3, next: 'day'   , prev:'year' },
+        day    : { subject: 'DAY(R.datetime_local)',    project: true,  level:4, next: 'hour'  , prev:'month'},
+        hour   : { subject: 'HOUR(R.datetime_local)',   project: true,  level:5, next: 'minute', prev:'day'  },
+        minute : { subject: 'MINUTE(R.datetime_local)', project: true,  level:6,                 prev:'hour' }
     },
     parseUrl: function(recording_url){
         var patternFound = false, resolved = false;
@@ -248,9 +249,9 @@ var Recordings = {
                         "SUBSTRING_INDEX(R.uri,'/',-1) as file, \n"+
                         "S.name as site, \n"+
                         "S.timezone, \n"+
-                        "S.legacy as legacy, \n"+
                         "R.uri, \n"+
                         "R.datetime, \n"+
+                        "R.datetime_local, \n"+
                         "R.mic, \n"+
                         "R.recorder, \n"+
                         "R.version, \n"+
@@ -325,6 +326,9 @@ var Recordings = {
                         return Recordings['__compute_' + property.replace(/-/g,'_')];
                     });
                 } else {
+                    data.forEach((d) => {
+                        d.legacy = Recordings.isLegacy(d)
+                    })
                     return data;
                 }
             }
@@ -435,11 +439,11 @@ var Recordings = {
 
     /**
      * Checks whether recording belongs to Arbimon (legacy) or RFCx platform
+     * Arbimon-based recordings are uploaded into buckets with similar naming (e.g. project_123, project_3, etc...)
      * @param {*} recording object containing the recording's data, like the ones returned in findByUrlMatch.
      */
-    isLegacy: async function(recording) {
-        const site = await this.getSiteModel(recording)
-        return site && site.legacy
+    isLegacy: function(recording) {
+        return recording.uri.startsWith('project_')
     },
 
     /** Downloads a recording from the bucket, storing it in a temporary file cache, and returns its path.
@@ -453,7 +457,7 @@ var Recordings = {
             if(!s3 || !s3RFCx){
                 defineS3Clients()
             }
-            const legacy = await this.isLegacy(recording)
+            const legacy = this.isLegacy(recording)
             let s3Client = legacy? s3 : s3RFCx
             const opts = {
                 Bucket : config(legacy? 'aws' : 'aws-rfcx').bucketName,
@@ -769,6 +773,14 @@ var Recordings = {
 
     },
 
+    calculateLocalTimeAsync: async function(site_id, datetime) {
+        var datetimeFormatted = moment.utc(datetime).format('YYYY-MM-DD HH:mm:ss');
+        return dbpool.query(`SELECT CONVERT_TZ('${datetimeFormatted}','UTC',S.timezone) as datetime_local FROM sites S WHERE S.site_id=${site_id}`, [])
+            .then((data) => {
+                return data[0].datetime_local;
+            })
+    },
+
     insert: function(recording, callback) {
 
         var schema = {
@@ -785,7 +797,8 @@ var Recordings = {
             file_size:       joi.number(),
             bit_rate:        joi.string(),
             sample_encoding: joi.string(),
-            upload_time:     joi.date()
+            upload_time:     joi.date(),
+            datetime_local:  joi.date(),
         };
 
         joi.validate(recording, schema, { stripUnknown: true }, function(err, rec) {
@@ -793,10 +806,10 @@ var Recordings = {
 
             queryHandler('INSERT INTO recordings (\n' +
                 '`site_id`, `uri`, `datetime`, `mic`, `recorder`, `version`, `sample_rate`, \n'+
-                '`precision`, `duration`, `samples`, `file_size`, `bit_rate`, `sample_encoding`, `upload_time`\n' +
+                '`precision`, `duration`, `samples`, `file_size`, `bit_rate`, `sample_encoding`, `upload_time`, `datetime_local`\n' +
             ') VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);', [
                 rec.site_id, rec.uri, rec.datetime, rec.mic || '(not specified)', rec.recorder || '(not specified)', rec.version || '(not specified)', rec.sample_rate,
-                rec.precision, rec.duration, rec.samples, rec.file_size, rec.bit_rate, rec.sample_encoding, rec.upload_time
+                rec.precision, rec.duration, rec.samples, rec.file_size, rec.bit_rate, rec.sample_encoding, rec.upload_time, rec.datetime_local
             ], callback);
         });
     },
@@ -881,17 +894,16 @@ var Recordings = {
         callback();
     },
     __compute_thumbnail_path_async : async function(recording){
-        if (recording.site_id && recording.legacy !== undefined && recording.site_external_id !== undefined) {
+        if (recording.site_id && recording.site_external_id !== undefined) {
             var site = {
                 site_id: recording.site_id,
-                legacy: !!recording.legacy,
                 external_id: recording.site_external_id
             }
         }
         else {
             var site = await Recordings.getSiteModel(recording)
         }
-        const legacy = site && site.legacy
+        const legacy = this.isLegacy(recording)
         if (legacy) {
             recording.thumbnail = 'https://' + config('aws').bucketName + '.s3.amazonaws.com/' + encodeURIComponent(recording.uri.replace(/\.([^.]*)$/, '.thumbnail.png'));
         }
@@ -1114,13 +1126,13 @@ var Recordings = {
                         for (let _1 of r) {
                             // Previously part of the query but now merged after
                             _1.site = siteData[_1.site_id].name;
-                            _1.legacy = siteData[_1.site_id].legacy;
+                            _1.legacy = Recordings.isLegacy(_1)
                             _1.site_external_id = siteData[_1.site_id].external_id;
                             _1.timezone = siteData[_1.site_id].timezone;
                             _1.imported = siteData[_1.site_id].project_id !== parameters.project_id;
                             Recordings.__compute_thumbnail_path_async(_1);
-                            if (!!siteData[_1.site_id].legacy !== true) {
-                                _1.file = `${moment.utc(_1.datetime).format('YYYY-MM-DD HH:mm:ss')}${path.extname(_1.file)}`;
+                            if (!_1.legacy) {
+                                _1.file = `${moment.utc(_1.datetime).format('YYYYMMDD_HHmmss')}${path.extname(_1.file)}`;
                             }
                         }
                     } else {