+++ scripts/db/009-recordings-table-add-datetime-local-column.sql
@@ -1,0 +1,6 @@
+ALTER TABLE `arbimon2`.`recordings`
+ADD COLUMN `datetime_local` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP;
+
+CREATE INDEX datetime_local_idx ON recordings(datetime_local);
+
+CREATE INDEX recordings_site_datetime_local_idx ON recordings(site_id, datetime_local);