+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Arbimon Release Notes
 
+## v3.0.21 - Apr 16, 2021
+
+New features:
+
+- PI-520 Modify the Arbimon upload page
+
+Resolved issues:
+- PI-714 Fix timezone issue for calendar on the Visualizer page
+- PI-731 Bounding box is no longer visible in the Pattern Matching for Citizen Scientist users
+- PI-732 User unable to create PM template
+- PI-730 Fix blank Visualizer page in Pattern Matching - Expert's View part in Citizen Scientist
+- Support for Core API v1.1.x changes
+- PI-739 The Visualizer page not loading before switching tabs
+- CE-537 Fix the location of the legend for the scatter plot
+- CE-538 Change the type of the request for getting recordings by aed ids
+
+Performance improvements:
+
+- CE-438 Do not request counts for proj where sites > 100 and recs > 100000
+- CE-451 Update code logic for determining recordings "legacy" attribute
+
 ## v3.0.20 - Mar 30, 2021
 
 New features: